More complete installation.

+++ edit-i7-lib/edit_install_extension.py
@@ -10,6 +10,9 @@ os.chdir(dir)
 
 if __name__ == '__main__':
     try:
+        p = os.path.join(inform7,"Extensions",sys.argv[2])
+	if not os.path.exists(p):
+		os.makedirs(p)
         p = os.path.join(inform7,"Extensions",sys.argv[2],sys.argv[1] + ".i7x")
         print p
         shutil.copyfile(sys.argv[3], p)
